requirements.txt
flask
selenium
python-dotenv
selenium-stealth
pandas
requests